Refactor rebuild executor, extract relationship parser, and apply code quality fixes
